Fixed a measurement glitch during range switching
When a range is switched, it may create a transient condition (dip or
spike). The ADC might have reported a value that had been measured over
two different ranges. Such a value is incorrect. This is now fixed and
measurements immediately following a range or mode (CC/CV) change are
discarded.